refactoring: use `pub(crate)` to drastically simplify the file layout
