--- src/internal/mod.rs
@@ -1,4 +1,0 @@
-pub mod scope;
-pub mod js;
-pub mod mem;
-pub mod vm;
